pyproject.toml
[tool.poetry]
name = "song-list"
version = "0.1.0"
description = ""
authors = ["vradchenko <vradchenko@callpanda.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
python-multipart = "^0.0.5"
aiohttp = {extras = ["speedups"], version = "^3.8.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
SQLAlchemy = "^1.4.37"
SQLAlchemy-Utils = "^0.38.2"
mysql-connector-python = "^8.0.29"

[tool.poetry.dev-dependencies]
pylint = "^2.14.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
line_length = 120

[tool.black]
line-length = 120